// backend/src/types/driver.types.ts
export interface CreateDriverDto {
  userId: string;
  licenseNumber?: string;
  nationality?: string;
  dateOfBirth?: string;
  bio?: string;
  emergencyContact?: string;
  medicalNotes?: string;
}

export interface UpdateDriverDto {
  licenseNumber?: string;
  nationality?: string;
  dateOfBirth?: string;
  bio?: string;
  emergencyContact?: string;
  medicalNotes?: string;
}